master: 	* Adding parameters to constructor 	* Fixing names 	* Adding tests for the LiveTestDynamoDBTable

+++ awstestutils/__init__.py
@@ -4,6 +4,8 @@ import json
 import logging
 import re
 
+import time
+
 log = logging.getLogger('awstestutils')
 
 TEST_NAME_PREFIX = 'test-'
@@ -303,15 +305,38 @@ class LiveTestDynamoDBTable(LiveTestBoto3Resource):
 
     Intended usage as a context manager:
 
-        >>> with LiveTestDynamoDBTable as table:
-        >>>     live.table.put_item(Item={
+        >>> with LiveTestDynamoDBTable() as table:
+        >>>     table.put_item(Item={
         >>>         'string_key': 'key1',
         >>>         'numeric_key': 0,
         >>>         'attribute_1': 'attribute'
         >>>     })
-        >>>     response = live.table.get_item(Key={
+        >>>     response = table.get_item(Key={
         >>>         'string_key': 'key1',
         >>>         'numeric_key': 0
+        >>>     })
+        >>>     print(response['Item'])
+
+    To customize the schema as a context manager you simply need to:
+
+        >>> key_schema, attributes_definitions, provisioned_throughput = LiveTestDynamoDBTable.create_key_schema(
+        >>>                                                                      partition_key_name='my_partition_key',
+        >>>                                                                      sorting_key_name='my_sorting_key',
+        >>>                                                                      partition_key_type='S',
+        >>>                                                                      sorting_key_type='N',
+        >>>                                                                      read_capacity_units=1,
+        >>>                                                                      write_capacity_units=1)
+        >>> with LiveTestDynamoDBTable(key_schema_definition=key_schema,
+        >>>                            attribute_definitions=attributes_definitions,
+        >>>                            provisioned_throughput=provisioned_throughput) as table:
+        >>>     table.put_item(Item={
+        >>>         'my_partition_key': 'key1',
+        >>>         'my_sorting_key': 0,
+        >>>         'attribute_1': 'attribute'
+        >>>     })
+        >>>     response = table.get_item(Key={
+        >>>         'my_partition_key': 'key1',
+        >>>         'my_sorting_key': 0
         >>>     })
         >>>     print(response['Item'])
     """
@@ -326,7 +351,7 @@ class LiveTestDynamoDBTable(LiveTestBoto3Resource):
         }
     ]
 
-    __DEFAULT_ATTRIBUTES_DEFINITIONS = [
+    __DEFAULT_ATTRIBUTE_DEFINITIONS = [
         {
             'AttributeName': 'string_key',
             'AttributeType': 'S'
@@ -355,7 +380,7 @@ class LiveTestDynamoDBTable(LiveTestBoto3Resource):
         :param sorting_key_type: Type for the tables's sorting key (String, Numeric, Set, etc)
         :param read_capacity_units: Quantity of read capacity units
         :param write_capacity_units: Quantity of write capacity units
-        :return: tuple with key_schema, attributes_definition and provisioned_throughput
+        :return: tuple with key_schema, attribute_definitions and provisioned_throughput
         """
         key_schema = []
         attributes_definitions = []
@@ -380,17 +405,25 @@ class LiveTestDynamoDBTable(LiveTestBoto3Resource):
             'WriteCapacityUnits': write_capacity_units
         }
 
-    def __init__(self, region_name=None):
+    def __init__(self, region_name=None,
+                 key_schema_definition=__DEFAULT_KEY_SCHEMA,
+                 attribute_definitions=__DEFAULT_ATTRIBUTE_DEFINITIONS,
+                 provisioned_throughput=__DEFAULT_PROVISIONED_THROUGHPUT):
         """
         Setup test manager.
 
         Assumes boto3 correctly configured
         :param region_name:
+        :param key_schema_definition:
+        :param attribute_definitions:
+        :param provisioned_throughput
         """
         self.table = None
         self.table_name = None
         self.dynamodb = boto3.resource('dynamodb', region_name=region_name)
-        pass
+        self.key_schema_definition = key_schema_definition
+        self.attribute_definitions = attribute_definitions
+        self.provisioned_throughput = provisioned_throughput
 
     def exists(self, table_name):
         for table in self.dynamodb.tables.all():
@@ -400,7 +433,7 @@ class LiveTestDynamoDBTable(LiveTestBoto3Resource):
 
     def create_table(self,
                      key_schema_definition=__DEFAULT_KEY_SCHEMA,
-                     attributes_definition=__DEFAULT_ATTRIBUTES_DEFINITIONS,
+                     attribute_definitions=__DEFAULT_ATTRIBUTE_DEFINITIONS,
                      provisioned_throughput=__DEFAULT_PROVISIONED_THROUGHPUT):
         """
         Creates the testing table with a name.
@@ -415,7 +448,7 @@ class LiveTestDynamoDBTable(LiveTestBoto3Resource):
         >>>         'KeyType': 'RANGE'
         >>>     }
         >>> ]
-        :param attributes_definition: The types for the table's key schema definition. By default uses:
+        :param attribute_definitions: The types for the table's key schema definition. By default uses:
         >>> [
         >>>     {
         >>>         'AttributeName': 'string_key',
@@ -438,21 +471,36 @@ class LiveTestDynamoDBTable(LiveTestBoto3Resource):
             table = self.dynamodb.create_table(
                 TableName=table_name,
                 KeySchema=key_schema_definition,
-                AttributesDefinition=attributes_definition,
+                AttributeDefinitions=attribute_definitions,
                 ProvisionedThroughput=provisioned_throughput)
         except Exception as e:
             raise RuntimeError('DynamoDB could not create table: %s' % e)
+        while table.table_status == 'CREATING':
+            time.sleep(0.01)
+            table = self.dynamodb.Table(table_name)
         self.table_name, self.table = table_name, table
 
     def destroy_table(self):
         """Destroys the created table."""
-        response = self.table.delete()
-        if self._is_error_call(response):
-            raise RuntimeError('DynamoDB coul not delete the table: %s' % response)
-        self.table, self.table_name = None, None
+        if self.table is None or self.table_name is None:
+            raise ValueError('inner table or table name are none')
+        while self.table.table_status == 'CREATING' or self.table.table_status == 'UPDATING':
+            time.sleep(0.01)
+            self.table = self.dynamodb.Table(self.table_name)
+        if self.table.table_status == 'ACTIVE':
+            response = self.table.delete()
+            if self._is_error_call(response):
+                raise RuntimeError('DynamoDB coul not delete the table: %s' % response)
+            self.table, self.table_name = None, None
+        elif self.table.table_status == 'DELETED':
+            pass
+        else:
+            raise ValueError('Unknown table state')
 
     def __enter__(self):
-        self.create_table()
+        self.create_table(key_schema_definition=self.key_schema_definition,
+                          attribute_definitions=self.attribute_definitions,
+                          provisioned_throughput=self.provisioned_throughput)
         return self.table
 
     def __exit__(self, *args):
